+++ lib/240325/data_source/user_api.dart
@@ -1,0 +1,15 @@
+import 'dart:convert';
+
+import 'package:http/http.dart' as http;
+
+import '../user.dart';
+
+class UserApi {
+  Future<List<User>> getUsers() async {
+    final response =
+        await http.get(Uri.parse('https://jsonplaceholder.typicode.com/users'));
+
+    List jsonList = jsonDecode(response.body);
+    return jsonList.map((e) => User.fromJson(e)).toList();
+  }
+}
