+++ lib/test/240325/todo_test.dart
@@ -1,7 +1,7 @@
 import 'package:test/test.dart';
 
 import '../../240325/todo.dart';
-import '../../연습/240325/data_source/todo_api.dart';
+import '../../240325/data_source/todo_api.dart';
 
 void main() {
   test('todo test', () async {
